+++ workspaces/quickstart/plugins/quickstart/src/components/QuickstartContent/QuickstartItem.tsx
@@ -28,6 +28,7 @@ import { QuickstartCtaLink } from './QuickstartCtaLink';
 import IconButton from '@mui/material/IconButton';
 import { QuickstartItemData } from '../../types';
 import { useTranslation } from '../../hooks/useTranslation';
+import { getTranslatedTextWithFallback } from '../../utils';
 
 export type QuickstartItemProps = {
   item: QuickstartItemData;
@@ -109,7 +110,7 @@ export const QuickstartItem = ({
           />
         </ListItemIcon>
         <ListItemText
-          primary={item.title}
+          primary={getTranslatedTextWithFallback(t, item.titleKey, item.title)}
           sx={{
             '& .v5-MuiTypography-root': {
               fontWeight: theme =>
@@ -150,7 +151,13 @@ export const QuickstartItem = ({
           }}
         >
           <ListItem>
-            <ListItemText primary={item.description} />
+            <ListItemText
+              primary={getTranslatedTextWithFallback(
+                t,
+                item.descriptionKey,
+                item.description,
+              )}
+            />
           </ListItem>
           <ListItem>
             <QuickstartCtaLink