+++ workspaces/quickstart/plugins/quickstart/src/index.ts
@@ -27,4 +27,7 @@ export * from './plugin';
 export { useQuickstartDrawerContext } from './hooks/useQuickstartDrawerContext';
 export type { QuickstartDrawerContextType } from './components/QuickstartDrawerContext';
 
-export { quickstartTranslations } from './translations';
+export {
+  quickstartTranslations,
+  quickstartTranslationRef,
+} from './translations';