+++ workspaces/translations/plugins/translations/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @red-hat-developer-hub/backstage-plugin-translations
 
+## 0.0.2
+
+### Patch Changes
+
+- a8787b0: Create an initial version of translations backend plugin
+  Add Download keys translations support in frontend plugin
+  Upgraded backstage to 1.42.5
+
 ## 0.0.1
 
 ### Patch Changes